Move RAG endpoints to router
